backend/requirements.txt
aiofiles==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.8" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.8" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.8" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.8" and python_version < "4.0"
exceptiongroup==1.1.2 ; python_version >= "3.8" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.8" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.8" and python_version < "4.0"
html5tagger==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.8" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.8" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.8" and python_version < "4.0"
httpx[http2]==0.24.1 ; python_version >= "3.8" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
jianshuresearchtools==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
lxml==4.9.3 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
pydantic-core==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
pydantic==2.0.2 ; python_version >= "3.8" and python_version < "4.0"
pymongo==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
sanic-routing==22.8.0 ; python_version >= "3.8" and python_version < "4.0"
sanic==23.3.0 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
sspeedup[ability-word-split,api-response-sanic,logging]==0.11.1 ; python_version >= "3.8" and python_version < "4.0"
tomd==0.1.3 ; python_version >= "3.8" and python_version < "4.0"
tracerite==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.8" and python_version < "4.0"
ujson==5.8.0 ; python_version >= "3.8" and python_version < "4.0" and sys_platform != "win32" and implementation_name == "cpython"
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython" and python_version >= "3.8" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.8" and python_version < "4.0"
